恢复resize功能，删除不必要的文件
