+++ easyui/components/panel.vue
@@ -100,8 +100,6 @@
        */
       bodyClass: String,
 
-      splitHandles: String,
-
       tools: Array
     },
 
@@ -115,9 +113,8 @@
         /**
          * Current height of the panel
          */
-        currentHeight: 0,
+        currentHeight: 0
 
-        isSplit: false
       }
     },
 
@@ -151,54 +148,29 @@
       getLayoutSize,
       setLayoutSize,
       move,
-      resize,
-      initSplit
+      resize
     },
 
     beforeMount() {
       // 需要在beforeMount节点执行，以确保能在子元素之前resize
       this.$nextTick(() => this.resize())
     },
+
     mounted() {
       // 仅根节点响应窗口resize事件
       if (this.$el === this.$root.$el) {
         LayoutEvents.$on('window-resize', () => this.$nextTick(() => this.resize()))
       }
+    },
 
-      this.initSplit()
-    },
     updated() {
-      this.initSplit()
     },
+
     beforeUpdate() {
       this.$nextTick(() => this.resize())
     },
+
     created() {
-    }
-  }
-
-  function initSplit() {
-    let el = $(this.$el),
-      vm = this
-
-    if (!this.isSplit && this.splitHandles) {
-      console.log('split', this.id, this.splitHandles)
-      el.resizable({
-        handles: this.splitHandles,
-        edge: 10,
-        onStartResize(e) {
-          vm.$emit('onStartResize', e)
-        },
-        onResize(e) {
-          vm.$emit('onResize', e)
-          return false
-        },
-        onStopResize(e) {
-          vm.$emit('onStopResize', e)
-          return false
-        }
-      })
-      this.isSplit = true
     }
   }
