--- easyui/directives/resizable/vue.resizable.js
@@ -1,12 +1,0 @@
-import $ from 'jquery'
-import Resizable from './jquery.resizable'
-
-Resizable()
-
-export default {
-  name: 'resizable',
-
-  bind (el) {
-    $(el).resizable()
-  }
-}
