+++ index.html
@@ -2,7 +2,6 @@
 <html lang="en">
   <head>
     <meta charset="utf-8">
-    <link rel="shortcut icon" href="favicon.ico">
     <title>vue-easyui</title>
 
     <style>
